fix(ai): address review findings on the Gemini agent
- Truncating a long thread could orphan a tool result from the assistant turn
  that opened it, leaving a `functionResponse` with no matching `functionCall`
  for Gemini to reject — a long-running chat would have stayed broken until
  the window moved past the boundary. `limitTranscript` now drops leading tool
  results after slicing, with regression cases.
- `verifyIdToken(token, true)`: the app ships `revokeAllSessions`, so without
  `checkRevoked` a signed-out-everywhere user kept streaming until expiry.
- Guard the SSE sink and break the stream loop when the client hangs up.
  Writing to a dead socket threw, and the throw escaped through the error path.
- Split `gemini-mapping.ts` (258 LOC) into `ag-ui-messages`, `gemini-schema`
  and `gemini-mapping`; every module is now well under the 250-LOC limit.
- Declare `@ag-ui/core` explicitly instead of relying on a hoisted transitive.
- `setup-staging.sh` provisions a GEMINI_API_KEY placeholder: `defineSecret`
  makes a missing secret fail `firebase deploy --only functions` outright, so
  even a staging project that never calls Gemini needs one to exist.
- Markdown fence language and given/when/then markers.
