+++ data-store/functions/src/ai/agent-prompt.ts
@@ -1,9 +1,6 @@
 /** Model backing the assistant — mirrors the motivation-quote generator. */
 export const AI_AGENT_MODEL = 'gemini-2.0-flash-lite';
 
-/** Upper bound on transcript turns forwarded to Gemini, newest kept. */
-export const MAX_TRANSCRIPT_MESSAGES = 40;
-
 export const AI_AGENT_SYSTEM_PROMPT = `You are the training coach inside "Pushup Tracker", a workout tracking web app.
 
 Behaviour:
@@ -20,12 +17,3 @@ Tools:
 - Use navigateTo when the user asks to see a page.
 - The app context lists every exercise id you may log, with its allowed rep
   range. Only use ids from that list.`;
-
-/** Keeps the prompt payload bounded on long threads. */
-export function limitTranscript<T>(
-  messages: readonly T[],
-  limit: number = MAX_TRANSCRIPT_MESSAGES
-): T[] {
-  if (messages.length <= limit) return [...messages];
-  return messages.slice(messages.length - limit);
-}
