+++ package.json
@@ -19,6 +19,7 @@
     "node": ">=24"
   },
   "devDependencies": {
+    "@ag-ui/core": "0.0.57",
     "@angular-devkit/core": "22.0.8",
     "@angular-devkit/schematics": "22.0.8",
     "@angular/build": "22.0.8",
