+++ web/src/app/ai/ai-assistant.config.spec.ts
@@ -16,24 +16,38 @@ describe('resolveAgentUrl', () => {
   });
 
   it('should collapse a whitespace-only value to the empty string', () => {
-    // given / when / then
-    expect(resolveAgentUrl({ agentUrl: '   ', agentId: 'default' })).toBe('');
+    // given
+    const config = { agentUrl: '   ', agentId: 'default' };
+
+    // when
+    const agentUrl = resolveAgentUrl(config);
+
+    // then
+    expect(agentUrl).toBe('');
   });
 });
 
 describe('isAiAssistantEnabled', () => {
   it('should be disabled when no agent URL is configured', () => {
-    // given / when / then
-    expect(isAiAssistantEnabled({ agentUrl: '', agentId: 'default' })).toBe(
-      false
-    );
+    // given
+    const config = { agentUrl: '', agentId: 'default' };
+
+    // when
+    const enabled = isAiAssistantEnabled(config);
+
+    // then
+    expect(enabled).toBe(false);
   });
 
   it('should be disabled when the agent URL is only whitespace', () => {
-    // given / when / then
-    expect(isAiAssistantEnabled({ agentUrl: '   ', agentId: 'default' })).toBe(
-      false
-    );
+    // given
+    const config = { agentUrl: '   ', agentId: 'default' };
+
+    // when
+    const enabled = isAiAssistantEnabled(config);
+
+    // then
+    expect(enabled).toBe(false);
   });
 
   it('should be enabled once an agent URL is configured', () => {
